requirements.txt
pandas
scikit-learn
pytorch-lightning
torchmetrics
albumentations